+++ cli/src/utils/target.ts
@@ -19,9 +19,12 @@ export const AVAILABLE_TARGETS = [
   'x86_64-unknown-freebsd',
   'i686-pc-windows-msvc',
   'armv7-unknown-linux-gnueabihf',
+  'armv7-unknown-linux-musleabihf',
   'armv7-linux-androideabi',
   'universal-apple-darwin',
   'riscv64gc-unknown-linux-gnu',
+  'powerpc64le-unknown-linux-gnu',
+  's390x-unknown-linux-gnu',
   'wasm32-wasi-preview1-threads',
 ] as const
 
@@ -37,6 +40,8 @@ export const DEFAULT_TARGETS = [
 export const TARGET_LINKER: Record<string, string> = {
   'aarch64-unknown-linux-musl': 'aarch64-linux-musl-gcc',
   'riscv64gc-unknown-linux-gnu': 'riscv64-linux-gnu-gcc',
+  'powerpc64le-unknown-linux-gnu': 'powerpc64le-linux-gnu-gcc',
+  's390x-unknown-linux-gnu': 's390x-linux-gnu-gcc',
 }
 
 // https://nodejs.org/api/process.html#process_process_arch
@@ -62,6 +67,7 @@ const CpuToNodeArch: Record<string, NodeJSArch> = {
   i686: 'ia32',
   armv7: 'arm',
   riscv64gc: 'riscv64',
+  powerpc64le: 'ppc64',
 }
 
 export const NodeArchToCpu: Record<string, string> = {
@@ -70,6 +76,7 @@ export const NodeArchToCpu: Record<string, string> = {
   ia32: 'i686',
   arm: 'armv7',
   riscv64: 'riscv64gc',
+  ppc64: 'powerpc64le',
 }
 
 const SysToNodePlatform: Record<string, Platform> = {
@@ -102,7 +109,11 @@ export interface Target {
  *   - `abi` = The ABI, for example `gnu`, `android`, `eabi`, etc.
  */
 export function parseTriple(rawTriple: string): Target {
-  if (rawTriple === 'wasm32-wasi-preview1-threads') {
+  if (
+    rawTriple === 'wasm32-wasi' ||
+    rawTriple === 'wasm32-wasi-preview1-threads' ||
+    rawTriple.startsWith('wasm32-wasip')
+  ) {
     return {
       triple: rawTriple,
       platformArchABI: 'wasm32-wasi',