+++ crates/backend/src/lib.rs
@@ -1,6 +1,9 @@
 #[macro_use]
 extern crate quote;
 
+use std::collections::HashMap;
+use std::sync::LazyLock;
+
 use proc_macro2::TokenStream;
 
 #[macro_use]
@@ -64,6 +67,7 @@ napi_ast_impl! {
  (Impl, NapiImpl),
  (Enum, NapiEnum),
  (Const, NapiConst),
+ (Type, NapiType),
 }
 
 pub(crate) static PRIMITIVE_TYPES: &[(&str, (&str, bool, bool))] = &[
@@ -80,6 +84,8 @@ pub(crate) static PRIMITIVE_TYPES: &[(&str, (&str, bool, bool))] = &[
   ("u8", ("number", false, false)),
   ("u16", ("number", false, false)),
   ("u32", ("number", false, false)),
+  // serde `Number`
+  ("Number", ("number", false, false)),
   ("u64", ("bigint", false, false)),
   ("i64n", ("bigint", false, false)),
   ("u128", ("bigint", false, false)),
@@ -92,6 +98,7 @@ pub(crate) static PRIMITIVE_TYPES: &[(&str, (&str, bool, bool))] = &[
   ("bool", ("boolean", false, false)),
   ("JsString", ("string", false, false)),
   ("String", ("string", false, false)),
+  ("RawCString", ("string", false, false)),
   ("str", ("string", false, false)),
   ("Latin1String", ("string", false, false)),
   ("Utf16String", ("string", false, false)),
@@ -103,3 +110,18 @@ pub(crate) static PRIMITIVE_TYPES: &[(&str, (&str, bool, bool))] = &[
   ("JsSymbol", ("symbol", false, false)),
   ("JsFunction", ("(...args: any[]) => any", true, false)),
 ];
+
+pub(crate) static TYPEDARRAY_SLICE_TYPES: LazyLock<HashMap<&str, &str>> = LazyLock::new(|| {
+  HashMap::from([
+    ("u8", "Uint8Array"),
+    ("i8", "Int8Array"),
+    ("u16", "Uint16Array"),
+    ("i16", "Int16Array"),
+    ("u32", "Uint32Array"),
+    ("i32", "Int32Array"),
+    ("f32", "Float32Array"),
+    ("f64", "Float64Array"),
+    ("u64", "BigUint64Array"),
+    ("i64", "BigInt64Array"),
+  ])
+});