+++ examples/napi-compat-mode/__tests__/function.spec.ts
@@ -36,9 +36,25 @@ test('should be able to create function from closure', (t) => {
   for (let i = 0; i < 100; i++) {
     t.is(
       bindings.testCreateFunctionFromClosure()(
-        ...Array.from({ length: i }).map((_, i) => i),
+        ...Array.from({ length: i }, (_, i) => i),
       ),
       `arguments length: ${i}`,
     )
   }
+})
+
+test('should be able to create nest function from closure', (t) => {
+  let callbackExecuted = false
+
+  const mockObject = {
+    on: (event: string, callback: Function) => {
+      t.is(event, 'on', 'Event name should be "on"')
+      callback()
+      callbackExecuted = true
+    },
+  }
+
+  const handle = bindings.testNestCreateFunctionFromClosure()
+  handle(mockObject)
+  t.true(callbackExecuted, 'Nested callback should have been executed')
 })